+++ dataconnect/buildSrc/src/main/kotlin/com/google/firebase/example/dataconnect/gradle/tasks/GenerateDataConnectSourcesTask.kt
@@ -14,8 +14,11 @@
  * limitations under the License.
  */
 
-package com.google.firebase.example.dataconnect.gradle
+package com.google.firebase.example.dataconnect.gradle.tasks
 
+import com.google.firebase.example.dataconnect.gradle.MyVariantProviders
+import com.google.firebase.example.dataconnect.gradle.runCommand
+import com.google.firebase.example.dataconnect.gradle.tweakConnectorYamlFiles
 import java.io.File
 import org.gradle.api.DefaultTask
 import org.gradle.api.file.DirectoryProperty
